Fix handling of audio tracks
Some streams have a different layout as the expected one, where the
video stream is not always the first one. This requires an explicit
mapping from the input to the output.

Regarding the capabilities of the target stream container the audio
track can be passed through or needs to be transcoded.

Introducing a new environment variable for how to handle the audio
track. It can be "auto" for automatic detection and handling,
"none" to remove any audio track, "silence" to add or replace the
audio track with silence, "aac" to force transcoding to AAC.